feat(file tool): merge file tool and rename tool
Instead of having separate file and rename tools, it makes more sense to unify them. This is done via a button at the top which can be toggled to either of "File Tool" or "Rename Tool."

Rename tool should effectively be deprecated, but kept in case it is still needed.